backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Redis
redis==5.0.1
hiredis==2.3.2

# ElasticSearch
elasticsearch[async]==8.11.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Rate Limiting
slowapi==0.1.9

# Background Tasks
celery==5.3.6
flower==2.0.1

# Email
fastapi-mail==1.4.1

# Validation & Utils
email-validator==2.1.0
python-dateutil==2.8.2

# Logging & Monitoring
python-json-logger==2.0.7
sentry-sdk[fastapi]==1.39.2
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0
faker==22.0.0

# Development
python-dotenv==1.0.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Other
python-slugify==8.0.1
bleach==6.1.0
pillow==10.2.0
